// src/types/StoreSearch.ts
export type StoreSearch = {
    address: string
    friday_reception_end_time: string
    friday_reception_from_time: string
    latitude: string
    longitude: string
    monday_reception_end_time: string
    monday_reception_from_time: string
    national_holiday_reception_end_time: string
    national_holiday_reception_from_time: string
    prefecture: string
    saturday_reception_end_time: string
    saturday_reception_from_time: string
    store_alias: string
    store_code: string
    store_image: string
    store_name: string
    sunday_reception_from_time: string
    thursday_reception_end_time: string
    thursday_reception_from_time: string
    tuesday_reception_end_time: string
    tuesday_reception_from_time: string
    wednesday_reception_end_time: string
    wednesday_reception_from_time: string
}